feat: aplicando factory method na criação de usuarios
