--- fctepodcast-backend/src/services/usuario/auth/logout.ts
@@ -1,9 +1,0 @@
-import { Request, Response } from "express";
-
-export const logout = async (req: Request, res: Response) => {
-  res.clearCookie("token").clearCookie("refreshToken").status(200).json({
-    status: "ok",
-    message: "Logout realizado com sucesso",
-    title: "Logout",
-  });
-};
